requirements.txt
azure-storage-blob
azure-core
python-dotenv
openai
PyPDF2
pytesseract
Pillow
pdf2image
pdfplumber
spacy